refactor(pie-monorepo-utils): DSW-000 reuse shared logic between /snapit \& /test-aperture
